move new entry into it's own file
